INT-10019: postgres: Base64 encode php-serialized data going into the DB
The mock api used in the mod_collaborate behat tests uses the php
serialize function and stores the result in the database. Since
serialize output generally contains zero byte characters, this fails on
systems using postgresql as the database.

https://bugs.php.net/bug.php?id=63344

The workaround for this implemented elsewhere in Moodle is to base64
encode/decode the serialized object on its way into and out of the
database.